+++ lib/spotify.dart
@@ -3,13 +3,14 @@
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:developer';
+import 'dart:math' as math;
 import 'dart:typed_data';
 
 import 'package:http/http.dart' as http;
 import 'package:oauth2/oauth2.dart' as oauth2;
 import 'package:spotify/spotify.dart';
 import 'package:meta/meta.dart';
-import 'dart:developer';
 
 import 'src/models/_models.dart';